requirements.txt
streamlit
openai
llama-parse
python-dotenv
nest_asyncio